% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_immunity.r
\name{adjust_immunity}
\alias{adjust_immunity}
\title{Estimate reproduction number from contact information and immunity profile}
\usage{
adjust_immunity(mixing, immunity)
}
\arguments{
\item{mixing}{A mixing matrix or set of mixing matrices, as returned by \code{\link{contact_matrix}}}

\item{immunity}{immunity profiles; this should be given as a vector of the same length as the number of rows/columns of the mixing matrix (or matrices); if given as a vector, each element of the vector should contain a value <1 representing the proportion of the population immune in the corresponding age group; any element set to "herd" will be set to 1-1/R0; if given as a data.frame, the number of rows need to be \code{n}, and every row will correspond to one of the samples}
}
\value{
a list contain vectors of adjusted immunities
}
\description{
This takes a contact survey to derive a contact matrix and rescales contacts to represent contacts with susceptibles. This is then combined with information on the basic reproduction number R0 to calculate the effective or net reproduction number.
}
\examples{
mixing <- socialmixr::contact_matrix(survey=socialmixr::polymod, age.limits = c(0, 5, 10))
adjust_immunity(mixing, immunity = c(0, 0.5, 0.8))
}
\author{
Sebastian Funk
}
