% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_immunity.r
\name{adjust_immunity}
\alias{adjust_immunity}
\title{Estimate reproduction number from contact information and immunity profile}
\usage{
adjust_immunity(survey, countries, immunity, n = 1, age.limits,
  model = c("age-specific", "homogeneous"), sample, ...)
}
\arguments{
\item{immunity}{immunity profiles; this should be given as a vector of the same length as \code{age.limits} or a data.frame; if given as a vector, each element of the vector should contain a value <1 representing the proportion of the population immune in the corresponding age group; any element set to "herd" will be set to 1-1/R0; if given as a data.frame, the number of rows need to be \code{n}, and every row will correspond to one of the samples}

\item{model}{"age-specific" (default) or "homogeneous"; if set to "homogeneous", the contact survey is not used to generate a contact matrix, but only for demographic information; beyond demography, random mixing is assumed.}

\item{...}{any further parameters for \code{\link{contact_matrix}}}
}
\value{
a list contain vectors of adjusted immunities
}
\description{
This takes a contact survey to derive a contact matrix and rescales contacts to represent contacts with susceptibles. This is then combined with information on the basic reproduction number R0 to calculate the effective or net reproduction number.
}
\author{
Sebastian Funk
}
