% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_age_dist.r
\name{epidemic_age_dist}
\alias{epidemic_age_dist}
\title{calculates the age distribution in an epidemic setting using the iterative method
of Wallinga (2006)}
\usage{
epidemic_age_dist(contact.matrix, q, child.mixing = NA, immunity = 0,
  final.size.start = 0.01, tol = 1e-05)
}
\arguments{
\item{contact.matrix:}{social contact matrix}

\item{q:}{probability of transmission per contact}

\item{child.mixing:}{first entry of contact matrix}

\item{immunity:}{proportion immune before the epidemic}

\item{final.size.start.start:}{starting value for inidence}

\item{tol:}{tolerance for stopping the iteration}
}
\value{
A vector of the final size (proportion of susceptibles infected) in each age group
}
\description{
calculates the age distribution in an epidemic setting using the iterative method
of Wallinga (2006)
}
